# docs/api_spec.yaml
openapi: 3.0.3
info:
  title: 商用地理信息虚假内容检查与反投喂平台 API
  description: |
    ## 概述
    本API提供地理信息虚假内容检测服务，支持单条/批量检测、规则管理、审核标注、统计报表等功能。
    
    ## 认证方式
    - **JWT Token**: 用于Web端用户认证，在Header中添加 `Authorization: Bearer <token>`
    - **API Key**: 用于服务端对接，在Header中添加 `X-API-Key: <api_key>`
    
    ## 请求限制
    - 基础版: 10,000次/天
    - 专业版: 50,000次/天
    - 企业版: 不限制
    
    ## 错误码
    | 错误码 | 说明 |
    |--------|------|
    | 400 | 请求参数错误 |
    | 401 | 未授权/Token过期 |
    | 403 | 权限不足 |
    | 404 | 资源不存在 |
    | 429 | 请求频率超限 |
    | 500 | 服务器内部错误 |
    
  version: 1.0.0
  contact:
    name: API Support
    email: api@geo-fake-detection.com
servers:
  - url: https://api.geo-fake-detection.com/v1
    description: 生产环境
  - url: https://api-staging.geo-fake-detection.com/v1
    description: 测试环境
...
paths:
...
  /records/{id}:
    get:
      tags: [Records]
      summary: 记录详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectionRecord'

    delete:
      tags: [Records]
      summary: 删除记录
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /rules/{id}:
    get:
      tags: [Rules]
      summary: 规则详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleConfig'

    put:
      tags: [Rules]
      summary: 更新规则
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleConfigUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleConfig'

    delete:
      tags: [Rules]
      summary: 删除规则
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /keywords/{id}:
    put:
      tags: [Keywords]
      summary: 更新关键词
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeywordUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Keyword'

    delete:
      tags: [Keywords]
      summary: 删除关键词
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /users/{id}:
    get:
      tags: [Users]
      summary: 用户详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    put:
      tags: [Users]
      summary: 更新用户
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    delete:
      tags: [Users]
      summary: 删除用户
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============================================
  # 系统管理
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
...
    UserUpdate:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
        status:
          type: string
...
    DetectionRecord:
      type: object
      properties:
        id:
          type: integer
        record_id:
          type: string
        device_id:
          type: string
        user_id:
          type: string
        timestamp:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        content_text:
          type: string
        content_type:
          type: string
        suspicion_score:
          type: number
        risk_level:
          type: string
        is_fake:
          type: boolean
        geo_score:
          type: number
        text_score:
          type: number
        simhash_score:
          type: number
        semantic_score:
          type: number
        reasons:
          type: array
          items:
            type: string
        review_status:
          type: string
        ground_truth:
          type: boolean
        created_at:
          type: string
          format: date-time
    RuleConfig:
      type: object
      properties:
        id:
          type: integer
        rule_type:
          type: string
          enum: [geo, text, simhash, semantic]
        rule_name:
          type: string
        rule_code:
          type: string
        rule_config:
          type: object
        description:
          type: string
        enabled:
          type: boolean
        priority:
          type: integer
        hit_count:
          type: integer
        created_at:
          type: string
          format: date-time
...
    RuleConfigUpdate:
      type: object
      properties:
        rule_name:
          type: string
        rule_config:
          type: object
        description:
          type: string
        enabled:
          type: boolean
        priority:
          type: integer
    Keyword:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        keyword:
          type: string
        weight:
          type: number
        risk_level:
          type: string
        source:
          type: string
        hit_count:
          type: integer
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    KeywordUpdate:
      type: object
      properties:
        weight:
          type: number
        risk_level:
          type: string
        enabled:
          type: boolean
...